refactor: remove unused React hooks and relocate likes buffer utilities
Removed useLikes, useLikesBuffer, LikeCountsSkeleton, and barrel export — all unused after Server Island migration.
Relocated buffer utilities from hooks/likes_buffer/internals/ to client/ to reflect their actual role as client-side modules.
Updated entry page to use server:defer with fallback skeleton and direct .astro import.
Updated E2E test to use Custom Element locator instead of astro-island hydration check.
