--- app/src/features/likes/components/LikeButton/index.ts
@@ -1,1 +1,0 @@
-export { LikeButton } from './LikeButton';
