--- app/src/features/likes/components/LikeCountsSkeleton/LikeCountsSkeleton.module.css
@@ -1,26 +1,0 @@
-.skeleton {
-  display: inline-flex;
-  width: 1rem;
-  height: 1rem;
-  background-color: var(--color-monotone-400);
-  border-radius: var(--border-radius);
-  animation: shimmer 1s infinite;
-}
-
-@keyframes shimmer {
-  0% {
-    background-color: var(--color-monotone-300);
-  }
-  50% {
-    background-color: var(--color-monotone-200);
-  }
-  100% {
-    background-color: var(--color-monotone-300);
-  }
-}
-
-@media (prefers-reduced-motion: reduce) {
-  .skeleton {
-    animation: none;
-  }
-}
