--- app/src/features/likes/components/LikeCountsSkeleton/LikeCountsSkeleton.tsx
@@ -1,5 +1,0 @@
-import styles from './LikeCountsSkeleton.module.css';
-
-export function LikeCountsSkeleton(): React.JSX.Element {
-  return <span className={styles.skeleton} />;
-}
