--- app/src/features/likes/components/LikeCountsSkeleton/index.ts
@@ -1,1 +1,0 @@
-export { LikeCountsSkeleton } from './LikeCountsSkeleton';
