--- app/src/features/likes/hooks/likes_buffer/index.ts
@@ -1,1 +1,0 @@
-export { useLikesBuffer } from './useLikesBuffer';
