+++ app/src/pages/entries/[id].astro
@@ -13,7 +13,7 @@ import Share from '../../features/entry/components/Share.astro';
 import Taxonomy from '../../features/entry/components/Taxonomy.astro';
 import { formatIsoString } from '../../features/entry/date';
 import { isFeatureEnabled } from '../../features/featureFlag';
-import { LikeButton } from '../../features/likes/components/LikeButton';
+import LikeButton from '../../features/likes/components/LikeButton/LikeButton.astro';
 import Body from '../../layouts/Body.astro';
 import GlobalFooter from '../../layouts/GlobalFooter.astro';
 import GlobalHeader from '../../layouts/GlobalHeader.astro';
@@ -136,6 +136,63 @@ const structuredData = JSON.stringify(
       flex-direction: column;
     }
   }
+
+  .LikeButtonFallback {
+    display: flex;
+    align-items: center;
+    gap: var(--space-component-1);
+    flex-direction: column;
+  }
+
+  .LikeButtonFallbackButton {
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    padding: var(--space-component-2);
+    border-radius: 50%;
+    background-color: transparent;
+    box-shadow: 0 1px 2px var(--color-monotone-300);
+    font-size: 1.5rem;
+    line-height: 1;
+  }
+  @media (prefers-color-scheme: dark) {
+    .LikeButtonFallbackButton {
+      background-color: var(--color-monotone-800);
+    }
+  }
+  :global([data-theme='light']) .LikeButtonFallbackButton {
+    background-color: transparent;
+  }
+  :global([data-theme='dark']) .LikeButtonFallbackButton {
+    background-color: var(--color-monotone-800);
+  }
+
+  .LikeButtonFallbackCount {
+    display: inline-flex;
+    width: 1rem;
+    height: 1rem;
+    background-color: var(--color-monotone-400);
+    border-radius: var(--border-radius);
+    animation: shimmer 1s infinite;
+  }
+
+  @keyframes shimmer {
+    0% {
+      background-color: var(--color-monotone-300);
+    }
+    50% {
+      background-color: var(--color-monotone-200);
+    }
+    100% {
+      background-color: var(--color-monotone-300);
+    }
+  }
+
+  @media (prefers-reduced-motion: reduce) {
+    .LikeButtonFallbackCount {
+      animation: none;
+    }
+  }
 </style>
 
 <Html>
@@ -167,11 +224,16 @@ const structuredData = JSON.stringify(
             {
               isLikeEnabled && (
                 <LikeButton
-                  client:visible
+                  server:defer
                   entryId={id}
                   entryTitle={entryTitle}
                   likeLabel={retrieveTranslation('entry.action.like')}
-                />
+                >
+                  <div slot="fallback" class="LikeButtonFallback">
+                    <span class="LikeButtonFallbackButton">👏</span>
+                    <span class="LikeButtonFallbackCount" />
+                  </div>
+                </LikeButton>
               )
             }
             <div class="Share">
